CoreOps: register refresh commands via public cache API
